rename images model to image, add static public backend folder, update creating image logic, update carousel shared module
