+++ express js/server.js
@@ -3,6 +3,8 @@ const compression = require('compression')
 const app = express()
 const port = 3000
 
+app.use(express.static(__dirname + '/public'));
+
 app.use(compression());
 // app.use(express.json());
 // app.use(express.urlencoded({
