+++ my-test-app/src/app/models/carousel.ts
@@ -1,14 +1,17 @@
+import { Image } from "./image";
+
 export class carousel{
     title: string;
     description: string;
-    imgPath: string;
-    imgAltText: string;
+    image: Image;
 
     constructor(obj?:carousel){
       this.title = obj?.title ?? '';
       this.description = obj?.description ?? '';
-      this.imgPath = obj?.imgPath ?? './assets/img/test1.jpg';
-      this.imgAltText = obj?.imgAltText ?? '';
+      this.image = obj?.image ?? new Image();
+
+      //Temporary solution
+      this.image.imagePath = 'backend/images/' + this.image.imagePath
     }
     
 }
